requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langdetect==1.0.9
MarkupSafe==3.0.2
nltk==3.9.1
openai==1.57.2
pydantic==2.10.3
pydantic_core==2.27.1
regex==2024.11.6
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
python-dotenv==1.0.0
vaderSentiment==3.3.2
pytest==7.4.0
gunicorn==20.1.0
